Unified errorCode, -stringForErrorCode: and networkError into one 'error' property.
The properties errorCode and networkError, and the method stringForErrorCode: are now all deprecated. They continue to function for now but may be removed in a future version. AS_NO_ERROR is also deprecated (see below).

For errorCode, you should now use stream.error.code or [[stream error] code].
For stringForErrorCode:, you should now use [stream.error localizedDescription] or [[stream error] localizedDescription].
For networkError, you should now use [stream.error localizedFailureReason] or [[stream error] localizedFailureReason]. This will soon be expanded to not be limited to network errors but the networkError property will remain limited.

The AudioStreamerErrorCode enum values have changed, with the exception of AS_NO_ERROR which is deprecated and now unused with the new error property. If you checked for errors by comparing the errorCode to not be AS_NO_ERROR or 0, you should now check if the error property is not nil. Alternatively, check the -doneReason method for AS_DONE_ERROR.